requirements.txt
PyYAML
kubernetes